@@ -1,30 +1,33 @@
 {
   "name": "javascript-menu",
-  "version": "1.0.0",
-  "description": "우아한테크코스 프리코스 점심 메뉴 추천 미션",
-  "dependencies": {
-    "@woowacourse/mission-utils": "^1.0.6"
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "start": "node src/index.js",
+    "test": "jest --detectOpenHandles"
   },
   "devDependencies": {
-    "jest": "^29.0.3"
+    "@babel/core": "^7.25.8",
+    "@babel/preset-env": "^7.25.8",
+    "babel-jest": "^29.6.0",
+    "jest": "^29.6.0"
   },
-  "scripts": {
-    "test": "jest --detectOpenHandles --watch"
+  "dependencies": {
+    "@woowacourse/mission-utils": "^2.2.0"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/woowacourse-precourse/javascript-menu.git"
+  "jest": {
+    "transform": {
+      "\\.js$": "babel-jest"
+    }
   },
-  "keywords": [],
-  "author": "woowacourse",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/woowacourse-precourse/javascript-menu/issues"
+  "babel": {
+    "presets": [
+      "@babel/preset-env"
+    ]
   },
-  "homepage": "https://github.com/woowacourse-precourse/javascript-menu#readme",
-  "engineStrict": true,
   "engines": {
-    "npm": ">=6.0.0",
-    "node": ">=14.0.0"
+    "npm": ">=10.8.2",
+    "node": ">=20.17.0"
   }
 }
